+++ src/acc/acc-nvim/lua/acc/downloader.lua
@@ -4,7 +4,7 @@
 local M = {}
 
 local ACC_VERSION = "0.3.2"
-local GITHUB_RELEASES_URL = "https://github.com/KeryxLabs/KeryxInstrumenta/releases/download"
+local GITHUB_RELEASES_URL = "https://github.com/EntasisLabs/instrumenta/releases/download"
 local ACC_RELEASE_TAG = "acc-engine/v" .. ACC_VERSION
 
 -- Returns the data directory where the binary is stored